Correct spacing with author and history in index
Was stripping too much whitespace with Jinja. So found that the
`authors` and `history` lines in the index where getting squished into
one line. Here we adjust the Jinja layout in the index file to ensure
the blank line is stripped if the author file is skipped, but no extra
stripping occurs.